fix bug with upload button and trash button
