+++ src/vue/text-editor/actions-menu-text-editor.vue
@@ -1,10 +1,6 @@
 <template>
     <div class="alpheios-alignment-editor-actions-menu">
       <div class="alpheios-alignment-editor-actions-menu__buttons">
-        <button class="alpheios-editor-button-tertiary alpheios-actions-menu-button" id="alpheios-actions-menu-button__upload"
-            @click="uploadTexts" :disabled="!docSourceEditAvailable" v-show="!onlyMetadata">
-            {{ l10n.getMsgS('ACTIONS_UPLOAD_TITLE') }}
-        </button>
         <button class="alpheios-editor-button-tertiary alpheios-actions-menu-button" id="alpheios-actions-menu-button__clear_text"
             @click="clearText" v-show="showClearText" :disabled="!docSourceEditAvailable">
             {{ l10n.getMsgS('ACTIONS_CLEAR_TEXT_TITLE') }}
@@ -14,25 +10,13 @@
             {{ toggleMetadataTitle }}
         </button>
       </div>
-      <div class="alpheios-alignment-editor-actions-menu__upload-block" v-show="showUploadBlock && docSourceEditAvailable" >
-          <input type="file" @change="loadTextFromFile" ref="fileupload">
-          <button class="alpheios-editor-button-tertiary alpheios-actions-menu-button" id="alpheios-actions-menu-button__metadata"
-              @click="showModal = true">
-              DTSAPI
-          </button>
-      </div>
-      <upload-dtsapi-block :showModal="showModal" @closeModal = "showModal = false" @uploadFromDTSAPI = "uploadFromDTSAPI"/>
     </div>
 </template>
 <script>
 import L10nSingleton from '@/lib/l10n/l10n-singleton.js'
-import UploadDTSAPIBlock from '@/vue/text-editor/upload-dtsapi-block.vue'
 
 export default {
   name: 'ActionsMenuTextEditor',
-  components: {
-    uploadDtsapiBlock: UploadDTSAPIBlock
-  },
   props: {
     textType: {
       type: String,
@@ -47,11 +31,6 @@ export default {
       required: false,
       default: false
     },
-    showUploadBlockFlag: {
-      type: Number,
-      required: false,
-      default: 1
-    },
     showClearTextFlag: {
       type: Number,
       required: false,
@@ -60,19 +39,11 @@ export default {
   },
   data () {
     return {
-      showUploadBlock: false,
       shownMetadataBlock: false,
-      showModal: false,
       showClearText: false
     }
   },
   watch: {
-    '$store.state.alignmentRestarted' () {
-      this.$refs.fileupload.value = ''
-    },
-    'showUploadBlockFlag' () {
-      this.uploadTexts()
-    },
     'showClearTextFlag' () {
       this.showClearText = true
     }
@@ -96,45 +67,9 @@ export default {
     }
   },
   methods: {
-    downloadSingle () {
-      this.$textC.downloadSingleSourceText(this.textType, this.textId)
-    },
-    /**
-     * Shows/Hides block with choose file input
-     */
-    uploadTexts () {
-      this.showUploadBlock = !this.showUploadBlock
-    },
-
     toggleMetadata () {
       this.shownMetadataBlock = !this.shownMetadataBlock
       this.$emit('toggle-metadata')
-    },
-
-    /**
-     * Creates FileReader and passes data from file to App component for parsing
-     */
-    loadTextFromFile(ev) {
-      const file = ev.target.files[0]     
-      if (!file) { return }
-      const extension = file.name.split('.').pop()
-
-      if (!this.$textC.checkUploadedFileByExtension(extension, false)) { return }
-
-      const reader = new FileReader()
-
-      reader.onload = e => {
-        this.$emit('upload-single', { text: e.target.result, extension })
-        this.showUploadBlock = false
-      }
-      reader.readAsText(file)
-
-      this.$refs.fileupload.value = ''
-    },
-
-    uploadFromDTSAPI (filedata) {
-      this.$emit('upload-single', { text: filedata.tei, lang: filedata.lang, extension: filedata.extension })
-      this.showUploadBlock = false
     },
 
     clearText () {
